move config and interface commands to commands dir, rename _plugins.py to plugin.py
