+++ src/nxscli/cli/main.py
@@ -6,12 +6,11 @@ from typing import Any
 import click
 from nxslib.proto.parse import Parser
 
-from nxscli.cli.cmd_interface import cmd_dummy, cmd_serial
 from nxscli.cli.environment import Environment, pass_environment
 from nxscli.iplugin import EPluginType, IPlugin
 from nxscli.logger import logger
 from nxscli.phandler import PluginHandler
-from nxscli.plugins_loader import configs_list, plugins_list
+from nxscli.plugins_loader import configs_list, interfaces_list, plugins_list
 
 ###############################################################################
 # Function: main
@@ -175,17 +174,16 @@ def cli_on_close(ctx: Environment) -> bool:
 def click_final_init() -> None:
     """Handle final Click initialization."""
     # interface commands
-    interfaces = [cmd_dummy, cmd_serial]
-    for intf in interfaces:
+    for intf in interfaces_list:
         main.add_command(intf)
 
     # configuration commands
-    for group in interfaces:
+    for group in interfaces_list:
         for cmd in configs_list:
             group.add_command(cmd)
 
     # plugin commands
-    for group in interfaces:
+    for group in interfaces_list:
         for plug in plugins_list:
             if plug.command:
                 group.add_command(plug.command)
