+++ src/nxscli/commands/cmd_interface.py
